Updates (vendored) bleach to 6.2.0 and removes dependencies on `six` and `html5lib` (#7013)
## Motivation for features / changes

Addresses #7010.

The old version of bleach that we were using, needed dependencies `six`
and `html5lib` to be installed, and the old version of "six" that we
were using for CI via bazel is not supported in python >= 3.12, which
causes some issues at least for local builds and CI when using newer
versions of python.

## Technical description of changes

Update the version of bleach specified in the `http_archive` rule used
by bazel and remove rules referring to `six` and `html5lib`.

## Detailed steps to verify changes work correctly (as executed by you)

Created a virtualenv with python 3.12+ and ran `bazel run //tensorboard
-- --logdir=<logdir>`, and verified that it runs correctly. Also CI
should be passing for this to get merged.