Updates version of urllib3 used by bazel (#7016)
## Motivation for features / changes

`urllib3` is a transitive dependency that "vendors" the `six` library in
it, but the version of `six` vendored within the version of urllib3 that
our code was using is not compatible with python >= 3.12.

This newer version is still within the major version number, but [after
1.26.5, which updated the vendored version of
`six`](https://github.com/urllib3/urllib3/blob/main/CHANGES.rst#1265-2021-05-26)
to one that runs with python >= 3.12.