Updates CI workflow to run on self-hosted runner (#7040)
## Motivation for features / changes
The standard runners provided by GitHub started running out of space, so
we set up self-hosted runners.

By internal policy, these runners cannot publish artifacts, e.g. to
PyPI, so the nightly workflows will need to be update to be produced by
separate infra, but at least this unblocks CI runs for PRs and such.

This change also removes the BUILD cache that was not actually used
before due to misconfiguration issue. To reduce maintenance burden
related to internal security policy and processes, the storage bucket
that was used for this purpose had public access removed, so we're
simply not using this anymore either way.

Note that this PR accomplishes the objective, which is to start running CI workflows in a self-hosted runner. However, tests seem to be failing in the new environment, and thus the CI workflow itself is failing.

I will proceed to merge this change, bypassing the requirement for the CI workflow to be successful, and we should address the test failures in one or more separate PRs.